Remove microformats gem dependency (#22923)
Looks like this gem was introduced as a dependency in 90aa720d2e for
testing Miroformat output.  The last test using the Microformats gem was
removed in 8ad51fba6c3887de311fb5fa0a2ae2cd8994c6d0, so I think it is
safe to remove this dependency.

For context, you [can't install the microformats gem with Ruby 3.2](https://github.com/microformats/microformats-ruby/pull/131),
so we can't currently bundle Mastodon with Ruby 3.2.  But since we don't
really need this gem, we can just remove it and unblock Ruby 3.2